+++ src/dodgr-to-sf.h
@@ -3,9 +3,9 @@
 
 #include <Rcpp.h>
 
-const float INFINITE_FLOAT =  std::numeric_limits<float>::max ();
-const double INFINITE_DOUBLE =  std::numeric_limits<double>::max ();
-const int INFINITE_INT =  std::numeric_limits<int>::max ();
+const float INFINITE_FLOAT =  std::numeric_limits <float>::max ();
+const double INFINITE_DOUBLE =  std::numeric_limits <double>::max ();
+const long int INFINITE_INT =  std::numeric_limits <long int>::max ();
 
 const std::string osm_p4s = "+proj=longlat +datum=WGS84 +no_defs";
