+++ src/heaps/bheap.cpp
@@ -13,7 +13,7 @@
 /* --- Constructor ---
  * Allocates and initialises a binary heap capable of holding n items.
  */
-BHeap::BHeap(unsigned int n)
+BHeap::BHeap(size_t n)
 {
     //    int i;
 
@@ -21,7 +21,7 @@ BHeap::BHeap(unsigned int n)
      * a[] since the indexing method does not use a[0].
      */
     a = new BHeapNode[n+1];
-    aPos = new unsigned int[n];
+    aPos = new size_t[n];
     //    for(i = 0; i <= n; i++) {
     //        a[i].item = 0;
     //        a[i].key = 0;
@@ -42,7 +42,7 @@ BHeap::~BHeap()
 /* --- min() ---
  * Returns the item with the minimum key in the heap.
  */
-unsigned int BHeap::min()
+size_t BHeap::min()
 {
     /* the item at the top of the binary heap has the minimum key value */
     return a[1].item;
@@ -56,13 +56,13 @@ double BHeap::getmin()
 /* --- insert() ---
  * Inserts an item $item$ with associated key value $key$ into the heap.
  */
-void BHeap::insert(unsigned int item, double key)
+void BHeap::insert(size_t item, double key)
 {
     /* i - insertion point
      * j - parent of i
      * y - parent's entry in the heap
      */
-    unsigned int i, j;
+    size_t i, j;
     BHeapNode y;
 
     /* $i$ initially indexes the new entry at the bottom of the heap */
@@ -94,13 +94,13 @@ void BHeap::insert(unsigned int item, double key)
 /* --- delete() ---
  * Deletes item $item$ from the heap.
  */
-void BHeap::deleteItem(unsigned int item)
+void BHeap::deleteItem(size_t item)
 {
     /* Decrease the number of entries in the heap and record the position of
      * the item to be deleted.
      */
-    const unsigned int n = --itemCount;
-    const unsigned int p = aPos[item];
+    const size_t n = --itemCount;
+    const size_t p = aPos[item];
 
     /* Heap needs adjusting if the position of the deleted item was not at the
      * end of the heap.
@@ -130,9 +130,9 @@ void BHeap::deleteItem(unsigned int item)
 /* --- decreaseKey() ---
  * Decreases the value of $item$'s key to the value $newKey$.
  */
-void BHeap::decreaseKey(unsigned int item, double newKey)
+void BHeap::decreaseKey(size_t item, double newKey)
 {
-    const unsigned int n = itemCount;
+    const size_t n = itemCount;
 
     itemCount = aPos[item] - 1;
     insert(item, newKey);
@@ -147,14 +147,14 @@ void BHeap::decreaseKey(unsigned int item, double newKey)
  * the root down, sifting up the minimum child until it is located in the
  * correct part of the binary heap.
  */
-void BHeap::siftUp(unsigned int p, unsigned int q)
+void BHeap::siftUp(size_t p, size_t q)
 {
     /* y - the heap entry of the root.
      * j - the current insertion point for the root.
      * k - the child of the insertion point.
      * z - heap entry of the child of the insertion point.
      */
-    unsigned int j, k;
+    size_t j, k;
     BHeapNode y, z;
 
     /* Get the value of the root and initialise the insertion point and child.
